+++ docs/Whats-New.md
@@ -510,6 +510,7 @@ Vanilla fixes:
 - Fixed MPDebug timer displaying when debug's visibility is off (by 11EJDE11)
 - Fixed the issue that units will goto farest location if target is closer than `MinimumRange` (by NetsuNegi)
 - Fixed the bug that techno with `Trainable=false` can be promoted if it's created by trigger event (by NetsuNegi)
+- Fixed the bug that ai will try to product aircraft even the airport has no free dock for it (by NetsuNegi)
 
 Phobos fixes:
 - Fixed the bug that `AllowAirstrike=no` cannot completely prevent air strikes from being launched against it (by NetsuNegi)